add the logic to create height - waist ratio
